--- src/main/java/com/api/QuickResponse/Controllers/RegisterMethod.java
@@ -1,15 +1,0 @@
-package com.api.QuickResponse.Controllers;
-
-import com.api.QuickResponse.Model.ItemRegister;
-import com.api.QuickResponse.Repository.QuickResponseRepository;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping(value = "/register")
-public class RegisterMethod {
-    @PostMapping
-    public Object register(@RequestBody ItemRegister itemRegister) {
-        QuickResponseRepository quickResponseRepository = QuickResponseRepository.getQuickResponseRepository();
-        return quickResponseRepository.register(itemRegister);
-    }
-}
