--- src/main/java/com/api/QuickResponse/Model/ItemLogin.java
@@ -1,43 +1,0 @@
-package com.api.QuickResponse.Model;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class ItemLogin {
-    @SerializedName("userName")
-    @Expose
-    private String userName;
-    @SerializedName("password")
-    @Expose
-    private String password;
-
-    public ItemLogin(String userName, String password) {
-        this.userName = userName;
-        this.password = password;
-    }
-
-    @Override
-    public String toString() {
-        return "ItemLogin{" +
-                "userName='" + userName + '\'' +
-                ", password='" + password + '\'' +
-                '}';
-    }
-
-    public String getUserName() {
-        return userName;
-    }
-
-    public void setUserName(String userName) {
-        this.userName = userName;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-}
