--- src/main/java/com/api/QuickResponse/Model/Manufacturing/EncodeBase64ToString.java
@@ -1,13 +1,0 @@
-package com.api.QuickResponse.Model.Manufacturing;
-
-import java.util.Base64;
-
-public class EncodeBase64ToString {
-    public EncodeBase64ToString() {
-    }
-
-    public static String base64Encode(String token) {
-        Base64.Encoder encode = Base64.getEncoder();
-        return encode.encodeToString(token.getBytes());
-    }
-}
