--- src/main/java/com/api/QuickResponse/Model/ReturnRegisterStatus/DataRegisterStatus.java
@@ -1,68 +1,0 @@
-package com.api.QuickResponse.Model.ReturnRegisterStatus;
-
-public class DataRegisterStatus {
-    public String userName;
-    public String fullName;
-    public int age;
-    public boolean gender;
-    public String id;
-
-    @Override
-    public String toString() {
-        return "DataRegisterStatus{" +
-                "userName='" + userName + '\'' +
-                ", fullName='" + fullName + '\'' +
-                ", age=" + age +
-                ", gender=" + gender +
-                ", id='" + id + '\'' +
-                '}';
-    }
-
-    public DataRegisterStatus(String userName, String fullName, int age, boolean gender, String id) {
-        this.userName = userName;
-        this.fullName = fullName;
-        this.age = age;
-        this.gender = gender;
-        this.id = id;
-    }
-
-    public String getUserName() {
-        return userName;
-    }
-
-    public void setUserName(String userName) {
-        this.userName = userName;
-    }
-
-    public String getFullName() {
-        return fullName;
-    }
-
-    public void setFullName(String fullName) {
-        this.fullName = fullName;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public boolean isGender() {
-        return gender;
-    }
-
-    public void setGender(boolean gender) {
-        this.gender = gender;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-}
