+++ src/main/java/com/api/QuickResponse/Model/UserLogin.java
@@ -1,0 +1,8 @@
+package com.api.QuickResponse.Model;
+
+public class UserLogin {
+    public String userName;
+    public String password;
+    public String accessToken;
+
+}
